requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
prettytable==2.1.0
pycryptodome==3.10.1
PySnooper==0.4.3
radius-eap-mschapv2-client==1.0.3
requests==2.25.1
shconfparser==2.2.2
six==1.15.0
tcping==0.1.1rc1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
